Reorganize business layer, change log type of cover image to update
