+++ implementation/src/main/java/database/dao/LogDAO.java
@@ -1,7 +1,7 @@
 package database.dao;
 
 import database.db.DBConnection;
-import repository.core.Log;
+import repository.core.pojo.Log;
 
 import java.sql.*;
 import java.time.LocalDateTime;
