+++ implementation/src/main/java/impl/ArtistManagerInMemory.java
@@ -1,15 +1,11 @@
 package impl;
 
-import repository.core.Artist;
-import repository.core.IArtistManager;
+import repository.core.interfaces.IManager;
+import repository.core.pojo.Artist;
 
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.Optional;
 import java.util.concurrent.CopyOnWriteArrayList;
-import java.util.stream.Collectors;
 
-public class ArtistManagerInMemory implements IArtistManager {
+public class ArtistManagerInMemory implements IManager {
     private CopyOnWriteArrayList<Artist> artists;
 
     public ArtistManagerInMemory() {
