+++ log-service/src/main/java/com/example/soap/service/LogEntry.java
@@ -1,21 +1,19 @@
 package com.example.soap.service;
 
-import org.glassfish.grizzly.http.util.TimeStamp;
-import repository.core.Log;
-import repository.core.LogFault;
+import repository.core.pojo.Log;
+import repository.core.exception.LogException;
 
 import javax.jws.WebMethod;
 import javax.jws.WebParam;
 import javax.jws.WebService;
 import javax.jws.soap.SOAPBinding;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 
 @WebService
 @SOAPBinding
 public interface LogEntry {
     @WebMethod(operationName = "listLog")
-    public ArrayList<Log> listLog(@WebParam(name="from") String from, @WebParam(name="to") String to, @WebParam(name="changeType") String changeType) throws LogFault;
+    public ArrayList<Log> listLog(@WebParam(name="from") String from, @WebParam(name="to") String to, @WebParam(name="changeType") String changeType) throws LogException;
     @WebMethod(operationName = "clearLog")
-    public String clearLog() throws LogFault;
+    public String clearLog() throws LogException;
 }
