+++ log-service/src/main/java/com/example/soap/service/LogEntryImpl.java
@@ -1,25 +1,23 @@
 package com.example.soap.service;
 
 import factories.ManagerFactory;
-import repository.core.ILogManager;
-import repository.core.Log;
-import repository.core.LogFault;
-import repository.core.RepException;
+import repository.core.interfaces.ILogManager;
+import repository.core.pojo.Log;
+import repository.core.exception.LogException;
+import repository.core.exception.RepException;
 
 import javax.jws.WebService;
-import java.text.ParseException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
-import java.util.IllegalFormatException;
 
 @WebService(endpointInterface = "com.example.soap.service.LogEntry")
 public class LogEntryImpl implements LogEntry {
     private ILogManager logManager = (ILogManager) ManagerFactory.LOG.getManager();
 
     @Override
-    public ArrayList<Log> listLog(String from, String to, String changeType) throws LogFault {
+    public ArrayList<Log> listLog(String from, String to, String changeType) throws LogException {
         ArrayList<Log> logs = new ArrayList<>();
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
         LocalDateTime fromDateTime = null;
@@ -33,7 +31,7 @@ public class LogEntryImpl implements LogEntry {
                 type = Log.ChangeType.valueOf(changeType);
             }
             else{
-                throw new LogFault("Enter a valid ChangeType (CASE SENSITIVE): ADD , UPDATE, DELETE");
+                throw new LogException("Enter a valid ChangeType (CASE SENSITIVE): ADD , UPDATE, DELETE");
             }
         }
 
@@ -45,7 +43,7 @@ public class LogEntryImpl implements LogEntry {
             }
             catch (DateTimeParseException pe)
             {
-                throw new LogFault("ERROR: Date format should be yyyy-MM-dd HH:mm:ss");
+                throw new LogException("ERROR: Date format should be yyyy-MM-dd HH:mm:ss");
             }
         }
         if(to != null && !to.equals("")){
@@ -54,7 +52,7 @@ public class LogEntryImpl implements LogEntry {
             }
             catch (DateTimeParseException pe)
             {
-                throw new LogFault("ERROR: Date format should be yyyy-MM-dd HH:mm:ss");
+                throw new LogException("ERROR: Date format should be yyyy-MM-dd HH:mm:ss");
             }
         }
 
@@ -79,13 +77,13 @@ public class LogEntryImpl implements LogEntry {
     }
 
     @Override
-    public String clearLog() throws LogFault {
+    public String clearLog() throws LogException {
         try {
             logManager.clearLog();
 
             return "Log cleared";
         } catch(RepException re) {
-            throw new LogFault(re.getMessage());
+            throw new LogException(re.getMessage());
         }
     }
 }
