--- source/src/main/java/repository/core/IArtistManager.java
@@ -1,13 +1,0 @@
-package repository.core;
-
-import java.util.ArrayList;
-
-public interface IArtistManager extends IManager{
-    /**
-    ArrayList<Artist> listArtist();
-    Artist getArtist(String nickname);
-    boolean addArtist(Artist artist);
-    boolean updateArtist(Artist artist);
-    boolean deleteArtist(String nickname);
-     */
-}
