--- source/src/main/java/repository/core/IManager.java
@@ -1,4 +1,0 @@
-package repository.core;
-
-public interface IManager {
-}
