--- source/src/main/java/repository/core/LogFault.java
@@ -1,10 +1,0 @@
-package repository.core;
-
-import javax.xml.ws.WebFault;
-
-@WebFault(name = "LogFault")
-public class LogFault extends Exception {
-    public LogFault(String message) {
-        super(message);
-    }
-}
