+++ source/src/main/java/repository/core/exception/RepException.java
@@ -1,4 +1,4 @@
-package repository.core;
+package repository.core.exception;
 
 public class RepException extends Exception {
     public RepException(String errorMessage) {
