+++ source/src/main/java/repository/core/interfaces/IManager.java
@@ -1,0 +1,4 @@
+package repository.core.interfaces;
+
+public interface IManager {
+}
