+++ source/src/main/java/repository/core/pojo/CoverImage.java
@@ -1,4 +1,4 @@
-package repository.core;
+package repository.core.pojo;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
